refactor: split shortform wizard state and actions
